# Conflicts:
#	roguetown.dme

+++ code/game/objects/structures/tables_racks.dm
@@ -101,6 +101,20 @@
 				user.stop_pulling()
 	return ..()
 
+/obj/structure/table/attack_right(mob/user)
+	var/obj/item/held = user.get_active_held_item()
+	var/obj/item/rogueweapon/bakers_peel/peel
+	if(istype(held, /obj/item/rogueweapon/bakers_peel))
+		peel = held
+		if(peel.unload_onto_table(src, user))
+			return TRUE
+	held = user.get_inactive_held_item()
+	if(istype(held, /obj/item/rogueweapon/bakers_peel))
+		peel = held
+		if(peel.unload_onto_table(src, user))
+			return TRUE
+	return ..()
+
 /obj/structure/table/proc/hideinside(mob/living/user)
 	var/sneak_level = user.get_skill_level(/datum/skill/misc/sneaking) || 0
 	var/sneaktime = max(10, 50 - (sneak_level * 10)) // Hard caps at 1 second at Expert and above.
@@ -158,6 +172,9 @@
 /obj/structure/table/proc/tablepush(mob/living/user, mob/living/pushed_mob)
 	if(HAS_TRAIT(user, TRAIT_PACIFISM))
 		to_chat(user, span_danger("Throwing [pushed_mob] onto the table might hurt them!"))
+		return
+	if(HAS_TRAIT(user, TRAIT_DEADITE)) //Deadites are too stupid to do this.
+		to_chat(user, span_warning("...what?"))
 		return
 	var/added_passtable = FALSE
 	if(!(pushed_mob.pass_flags & PASSTABLE))
@@ -554,10 +571,16 @@
 	. += span_blue("Right-Click to fold the table.")
 
 /obj/structure/table/wood/folding/attack_right(mob/user)
+	if(..())
+		return TRUE
 	user.visible_message(span_notice("[user] folds [src]."), span_notice("You fold [src]."))
+	//Caustic Edit - Stop the folding table from deleting hidden people!
+	if(hiddenguy)
+		src.unhide(hiddenguy)
+	//Caustic Edit End
 	new /obj/item/folding_table_stored(drop_location())
 	qdel(src)
-	return ..()
+	return TRUE
 
 /*
  * Racks
@@ -646,11 +669,32 @@
 	climb_offset = 0
 	pixel_y = 32
 
+	var/move_on_craft = TRUE // Caustic Edit
+
+// Caustic Edit
+/obj/structure/rack/rogue/shelf/OnCrafted(dirin)
+	if(move_on_craft) 
+		pixel_x = 0
+		pixel_y = 0
+		switch(dirin)
+			if(NORTH)
+				pixel_y = 32
+			if(SOUTH)
+				pixel_y = -32
+			if(EAST)
+				pixel_x = 32
+			if(WEST)
+				pixel_x = -32
+	. = ..()
+// Caustic Edit end
+
 /obj/structure/rack/rogue/shelf/big
 	icon = 'icons/roguetown/misc/structure.dmi'
 	icon_state = "shelf_big"
 	dir = SOUTH
 	pixel_y = 16
+
+	move_on_craft = FALSE // Caustic Edit
 
 /obj/structure/rack/rogue/shelf/biggest
 	icon_state = "shelf_biggest"
@@ -658,6 +702,8 @@
 	// This isn't pixel-shifted, and therefore should be dense.
 	density = TRUE
 	climb_offset = 10
+
+	move_on_craft = FALSE // Caustic Edit
 
 /obj/structure/rack/rogue/shelf/notdense // makes the wall mounted one less weird in a way, got downside of offset when loaded again tho
 	density = FALSE
